django/requirements.txt
Django==2.2.13
requests
beautifulsoup4
django-crispy-forms
django-bootstrap4
django-bootstrap-datepicker-plus
django-allauth
mysqlclient
uwsgi

autopep8
flake8
black
factory_boy
django-debug-toolbar
